refactor: Move NewTestContext from test_helpers.go to environment.go
NewTestContext is not just a test helper - it's used in production code
(e2e_generator.go) to create contexts for handler execution. Therefore,
it belongs in environment.go alongside other context-related functions
like SendTo, Goto, and Halt.

Changes:
- Moved NewTestContext from test_helpers.go to environment.go
- Removed unused context import from test_helpers.go
- Enhanced documentation to clarify its use for handler execution
- No functional changes

All tests pass.
